+++ noodles-cram/CHANGELOG.md
@@ -44,6 +44,9 @@
 
   * cram/data_container/slice/builder: Detach all records before writing.
 
+  * cram/record/convert: Create read group (`RG`) data field if a read group ID
+    is set.
+
   * cram/writer/record: Fix possible wrapping when casting when writing number
     of read features.